Add files via upload
The book indexer in Python is a powerful tool for organizing and categorizing books. It automatically scans book content, extracts relevant keywords, and creates an index for efficient searching. With its intuitive interface and customizable features, it simplifies the process of finding specific information within a large collection of books.
[You can know about the task from the assignment.txt  . and  VS file contains the python source code and Ans  files contain the output]